Extract SQS consumer and SNS pub to the Product Service
