+++ product-service/handler.js
@@ -1,4 +1,5 @@
 import * as database from './database';
+import catalogBatchProcessHandler from './handlers/catalogBatchProcess';
 import createProductHandler from './handlers/createProduct';
 import getProductByIdHandler from './handlers/getProductById';
 import getProductsListHandler from './handlers/getProductsList';
@@ -7,6 +8,8 @@ import wrapHttpHandler from './http/wrapHttpHandler';
 /**
  * Index of the functions.
  */
+export const catalogBatchProcess = wrapHttpHandler(catalogBatchProcessHandler);
+
 export const createProduct = wrapHttpHandler(createProductHandler(database));
 
 export const getProductById = wrapHttpHandler(getProductByIdHandler(database));
